refactor(auth): move email OTP routes
