+++ docs/context/architecture/composition.md
@@ -6,6 +6,7 @@ sources:
   - src/treg/domain/identity/mcp_oauth.py
   - src/treg/domain/identity/session.py
   - src/treg/routers/admin.py
+  - src/treg/routers/auth.py
   - src/treg/routers/web.py
   - scripts/dump_surface.py
 related:
@@ -18,9 +19,9 @@ related:
 # Application composition
 
 `bootstrap.create_app(role)` is the FastAPI composition root. `api.py` hosts the ordered route table;
-the Catalog, web, and admin modules define concern-specific `APIRouter` blocks that `api.py` appends
-at their legacy registration points. It then calls the factory once at EOF so the deployed and
-documented `treg.api:app` import path remains the default `all` role.
+the auth, Catalog, web, and admin modules define concern-specific `APIRouter` blocks that `api.py`
+appends at their legacy registration points. It then calls the factory once at EOF so the deployed
+and documented `treg.api:app` import path remains the default `all` role.
 
 The factory owns concrete assembly: the three pure-ASGI middleware registrations, five exception handlers,
 static mounts, optional MCP mount and lifespan, GET-to-HEAD widening, the OpenAPI wrapper that hides
