+++ docs/context/interface/api.md
@@ -6,6 +6,7 @@ sources:
   - src/treg/api.py
   - src/treg/routers/__init__.py
   - src/treg/routers/admin.py
+  - src/treg/routers/auth.py
   - src/treg/routers/auth_helpers.py
   - src/treg/routers/catalog.py
   - src/treg/routers/signup_cookies.py
@@ -27,11 +28,11 @@ related:
 
 # The API
 
-Route definitions live on `api.router`; the open Catalog JSON block is defined in
-`routers.catalog`, the three presentation blocks are defined in `routers.web`, and the two
-cross-tenant admin read/report blocks are defined in `routers.admin`. `api.py` attaches each block at
-its original registration point. `bootstrap.create_app()` assembles the combined route table into
-FastAPI roles.
+Route definitions live on `api.router`; the email OTP block is defined in `routers.auth`, the open
+Catalog JSON block is defined in `routers.catalog`, the three presentation blocks are defined in
+`routers.web`, and the two cross-tenant admin read/report blocks are defined in `routers.admin`.
+`api.py` attaches each block at its original registration point. `bootstrap.create_app()` assembles
+the combined route table into FastAPI roles.
 `api.app` remains the deployed, backward-compatible `all` role. Everything the CLI + skill do is one
 HTTP call over this. The factory lifespan
 runs `init_db()`, then `_backfill_provider_extra_tools()` (the idempotent repair for provider registry
